+++ source/SemVerAnalyzer/AppModule.cs
@@ -1,9 +1,8 @@
-using System.Linq;
-
 using Autofac;
 
 using Pushpay.SemVerAnalyzer.Engine;
 using Pushpay.SemVerAnalyzer.Nuget;
+using SemVerAnalyzer.Abstractions;
 
 namespace Pushpay.SemVerAnalyzer
 {
@@ -24,7 +23,7 @@ namespace Pushpay.SemVerAnalyzer
 				.AsSelf();
 
 			builder.RegisterAssemblyTypes(ThisAssembly)
-				.Where(t => t.IsAssignableTo<IVersionAnalysisRule>() && !_settings.DisabledRules.Contains(t.Name))
+				.Where(t => t.IsAssignableTo<IVersionAnalysisRule>() && _settings.GetOverrideType(t) != RuleOverrideType.Ignore)
 				.AsImplementedInterfaces()
 				.AsSelf();