+++ source/SemVerAnalyzer/Engine/Rules/ReferencesMajorBumpedRule.cs
@@ -1,4 +1,5 @@
 using Pushpay.SemVerAnalyzer.Assembly;
+using SemVerAnalyzer.Abstractions;
 
 namespace Pushpay.SemVerAnalyzer.Engine.Rules
 {