+++ source/SemVerAnalyzer/Engine/Rules/TypeAddedRule.cs
@@ -1,4 +1,5 @@
 using dnlib.DotNet;
+using SemVerAnalyzer.Abstractions;
 
 namespace Pushpay.SemVerAnalyzer.Engine.Rules
 {